Only prompt to save config if config wasn't passed

+++ src/main.js
@@ -3,7 +3,7 @@ import fs from 'fs'
 import chalk from 'chalk'
 import inquirer from 'inquirer'
 import { kickstart } from './core/index.js'
-import { originalConfig } from './helpers/globals.js'
+import { args, originalConfig } from './helpers/globals.js'
 import { clearUndefined, formatTime, parseArgs } from './helpers/index.js'
 import { logSuccess } from './helpers/logger.js'
 import { promptQuestions } from './helpers/prompt.js'
@@ -18,14 +18,16 @@ const main = async () => {
   await kickstart()
   clearUndefined()
 
-  const { answer } = await inquirer.prompt({
-    type: 'confirm',
-    prefix: `${chalk.magenta('?')}`,
-    name: 'answer',
-    message: 'Would you like to save your config file? (to reuse in the future)'
-  })
+  if (!args.config) {
+    const { answer } = await inquirer.prompt({
+      type: 'confirm',
+      prefix: `${chalk.magenta('?')}`,
+      name: 'answer',
+      message: 'Would you like to save your config file? (to reuse in the future)'
+    })
 
-  if (answer) fs.writeFileSync('kickstart.config.json', JSON.stringify(originalConfig, null, 2))
+    if (answer) fs.writeFileSync('kickstart.config.json', JSON.stringify(originalConfig, null, 2))
+  }
 
   const executionEnd = performance.now()
   logSuccess('------', false)
